RAG_model/requirements.txt
google-generativeai>=0.8.0
cohere>=5.0.0
chromadb>=0.4.18
tqdm>=4.66.0
numpy>=1.24.0
PyPDF2>=3.0.0
